got rid of 10-k signal lag, did some cleaning, added utils for ten k backfill
